restructure python sdk lib to have pyo bindings in sep. modules
