+++ crates/prime-protocol-py/src/orchestrator/mod.rs
@@ -1,0 +1,19 @@
+use pyo3::prelude::*;
+
+/// Prime Protocol Orchestrator Client - for managing and distributing tasks
+#[pyclass]
+pub struct OrchestratorClient {
+    // TODO: Implement orchestrator-specific functionality
+}
+
+#[pymethods]
+impl OrchestratorClient {
+    #[new]
+    #[pyo3(signature = (rpc_url, private_key=None))]
+    pub fn new(rpc_url: String, private_key: Option<String>) -> PyResult<Self> {
+        // TODO: Implement orchestrator initialization
+        let _ = rpc_url;
+        let _ = private_key;
+        Ok(Self {})
+    }
+}
