+++ crates/prime-protocol-py/src/validator/mod.rs
@@ -1,0 +1,19 @@
+use pyo3::prelude::*;
+
+/// Prime Protocol Validator Client - for validating task results
+#[pyclass]
+pub(crate) struct ValidatorClient {
+    // TODO: Implement validator-specific functionality
+}
+
+#[pymethods]
+impl ValidatorClient {
+    #[new]
+    #[pyo3(signature = (rpc_url, private_key=None))]
+    pub fn new(rpc_url: String, private_key: Option<String>) -> PyResult<Self> {
+        // TODO: Implement validator initialization
+        let _ = rpc_url;
+        let _ = private_key;
+        Ok(Self {})
+    }
+}
